feat: add context_noise parameter to CleanKVCacheBlock for enhanced control over KV cache updates
Introduced a new input parameter, context_noise, to the CleanKVCacheBlock class, allowing users to specify the noise level during KV cache updates. This addition improves flexibility in managing the denoising process. Additionally, streamlined import statements for better readability in clean_kv_cache.py. Updated the context timestep calculation to utilize the context_noise parameter.

+++ src/scope/core/pipelines/wan2_1/blocks/set_timesteps.py
@@ -99,7 +99,7 @@ class SetTimestepsBlock(ModularPipelineBlocks):
             indices = (1000 - denoising_step_list).clamp(
                 0, len(timesteps_with_zero) - 1
             )
-            denoising_step_list = timesteps_with_zero[indices].long()
+            denoising_step_list = timesteps_with_zero[indices]
 
         if block_state.current_denoising_step_list is None or not torch.equal(
             block_state.current_denoising_step_list, denoising_step_list
